% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{affected_pathogens}
\alias{affected_pathogens}
\title{outputs to fade out pathongens not applicable to country.}
\format{
A data frame with 249 rows and 4 variables:
\describe{
\item{iso3}{Country ISO3 codes.}
\item{yellow_fever}{TRUE if country is affected by yellow fever.}
\item{cholera}{TRUE if country is affected by cholera.}
\item{meningitis}{TRUE if country is affected by meningitis.}
}
}
\usage{
affected_pathogens
}
\description{
outputs to fade out pathongens not applicable to country.
}
\keyword{datasets}
