% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_billions_example}
\alias{all_billions_example}
\title{Example dataset of all three billions}
\format{
A data frame with 3244 rows and 13 variables:
\describe{
\item{iso3}{Country ISO3 codes.}
\item{year}{Year.}
\item{ind}{Indicator.}
\item{value}{Value.}
\item{use_dash}{Bolean indicating.}
\item{source}{Source of the data.}
\item{type}{Type of data.}
\item{transform_value}{Transformed value by billionaiRe package.}
\item{level}{Levels calculated by billionaiRe package.}
\item{contribution}{Contribution of indicator to billion.}
\item{contribution_percent}{Percent contribution to billion.}
\item{population}{Population numbers to be used.}
\item{contribution_percent_total_pop}{Percent of total population contribution to billion.}
}
}
\usage{
all_billions_example
}
\description{
This dataset contains dummy data for all three billions. The data is provided
as an example and does not reflect the situation in any of the the
three countries. This dataset should be understood as having going through
the \code{billionaiRe} pipeline for all three billions. See the
\href{https://github.com/gpw13/billionaiRe}{billionaiRe} package for more details.
}
\keyword{datasets}
