% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_exports.R
\name{assert_valid_BorderStyle}
\alias{assert_valid_BorderStyle}
\title{Validate BorderStyle}
\usage{
assert_valid_BorderStyle(borderStyle)
}
\arguments{
\item{borderStyle}{borderStyle parameter as in openxlsx::createStyle}
}
\description{
Validate BorderStyle
}
\author{
Philipp Schauberger
}
