% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_exports.R
\name{assert_valid_Colour}
\alias{assert_valid_Colour}
\title{Validate colour inputed}
\usage{
assert_valid_Colour(colour, errorMsg = "Invalid colour")
}
\arguments{
\item{colour}{colour as \code{colours()}}

\item{errorMsg}{error message to be passed.}
}
\description{
Validate colour inputed
}
\author{
Philipp Schauberger
}
