% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_country_summary_xls}
\alias{export_country_summary_xls}
\title{Export country summary to Excel}
\usage{
export_country_summary_xls(
  df,
  iso,
  billion = c("hpop", "hep", "uhc", "all"),
  value_col = "value",
  transform_value_col = "transform_value",
  scenario_col = NULL,
  contribution = "contribution",
  contribution_pct = paste0(contribution, "_percent"),
  contribution_pct_total_pop = paste0(contribution, "_percent_total_pop"),
  default_scenario = "default",
  start_year = 2018,
  end_year = 2019:2025,
  output_folder = "outputs",
  version = whdh::get_formatted_timestamp()
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso}{ISO3 code of country to summarize.}

\item{billion}{Billion indicator names to return, either "hep", "hpop", "uhc"
, or "all".}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value_col}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value_col}.}

\item{contribution_pct_total_pop}{Column name of column(s) to store contribution
(percent of total population of the country) values. Must be the same length
as \code{transform_value_col}.}

\item{default_scenario}{name of the default scenario.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to
2025.}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{version}{character vector identifying the version of the plot to be passed.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_country_summary_xls} Export a country-specific for all three
billions or for a specific billion.
}
