% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot_comparison.R
\name{export_plot_comparison_pdf}
\alias{export_plot_comparison_pdf}
\title{Export plot comparison of two billionaiRe data frames}
\usage{
export_plot_comparison_pdf(
  ...,
  indicator,
  output_folder = "outputs",
  run_name = NULL,
  ind = "ind",
  scenario = "scenario",
  default_scenario = "default",
  type_col = "type",
  value = "value",
  start_year = 2018,
  year = "year",
  iso3 = "iso3",
  ind_ids = billionaiRe::billion_ind_codes("all", include_calculated = TRUE)
)
}
\arguments{
\item{...}{Name-value pair of data frames to use e.g. (new = df_1).The name
(if provided) gives the name of the data frame and the value is the data
frame itself. The name is used to appear in the legend of the plot if
provided.}

\item{indicator}{name of indicator to plot}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{run_name}{name of the run of plots. Default to NULL. If used, a new
folder will be added to \code{output_folder} to store the results.}

\item{ind}{Column name of column with indicator names.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{type_col}{Column name of column with type data.}

\item{value}{Column name of column with indicator values.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{group_iso3}{if TRUE (default), iso3s will be grouped by 54 to have more readable graphs.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_plot_comparison_pdf} exports plots to pdf comparing two billionaiRe
data frames.
}
