% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_indicator.R
\name{export_plot_timeseries_indicator}
\alias{export_plot_timeseries_indicator}
\title{Export plot timeseries for an indicator}
\usage{
export_plot_timeseries_indicator(
  df,
  indicator,
  format = c("pdf", "png", "plotly"),
  output_folder = "outputs",
  ind = "ind",
  scenario = NULL,
  default_scenario = "default",
  type_col = "type",
  transform_value = "transform_value",
  year = "year",
  iso3 = "iso3",
  ind_ids = billionaiRe::billion_ind_codes("all", include_calculated = TRUE)
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{indicator}{name of indicator to plot}

\item{format}{format of the export to produced. Can be either
\itemize{
\item \strong{pdf}
\item \strong{png} for images
\item \strong{plotly} for html interactive graphs
}}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{ind}{Column name of column with indicator names.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{type_col}{Column name of column with type data.}

\item{transform_value}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{year}{Column name of column with years.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_plot_timeseries_indicator} exports plots in the \code{format} specified.
}
