% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot_timeseries.R
\name{export_plot_timeseries_indicator_iso3_group}
\alias{export_plot_timeseries_indicator_iso3_group}
\title{Export timeseries for a group of countries}
\usage{
export_plot_timeseries_indicator_iso3_group(
  df,
  iso3_group,
  indicator,
  output_folder,
  scale,
  value_col,
  default_scenario,
  start_year,
  experiment
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{iso3_group}{named list of iso3 to be ploted}

\item{indicator}{name of indicator to plot}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{scale}{type of scale to be exported. Can be either:
\itemize{
\item \code{common}: same scale for all iso3
\item \code{free}: individual scale for each iso3
\item \code{combined}: combination of common and free.
By default, \code{combined} is used
}}

\item{value_col}{Column name of column with indicator values.}

\item{default_scenario}{name of the default scenario.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{experiment}{name of the experiment of plots. Default to NULL. If used, a new
folder will be added to \code{output_folder} to store the results.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_plot_timeseries_indicator_pdf} exports plots to pdf.
}
