% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_summary.R
\name{export_uhc_country_summary_xls}
\alias{export_uhc_country_summary_xls}
\title{Export country summary to Excel for UHC billion
\code{export_uhc_country_summary_xls} Export a country-specific for UHC billion.}
\usage{
export_uhc_country_summary_xls(
  df,
  wb,
  iso,
  value_col = "value",
  transform_value_col = "transform_value",
  scenario_col = NULL,
  contribution = "contribution",
  default_scenario = "default",
  start_year = 2018,
  end_year = 2019:2025,
  scenario_end_year = end_year,
  sheet_prefix = "UHC",
  output_folder = "outputs",
  ind_ids = billionaiRe::billion_ind_codes("uhc", include_calculated = TRUE)
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{iso}{ISO3 code of country to summarize.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value_col}.}

\item{default_scenario}{name of the default scenario.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to
2025.}

\item{scenario_end_year}{End year of the scenarios. Defaults to \code{end_year}.}

\item{sheet_prefix}{Character prefix to add in front of export sheets}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_hep_country_summary_xls} Export a country-specific for HEP billion.
}
