% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{get_baseline_projection_df}
\alias{get_baseline_projection_df}
\title{Get baseline and projections data frame for specified dates}
\usage{
get_baseline_projection_df(
  df,
  value_col,
  transform_value_col,
  start_year,
  end_year,
  ind_df
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to
2025.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{get_latest_reported()} gets the latest reported data available for estimated
or reported data. Used in write functions.
}
