% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_bounds_regional.R
\name{get_box_bounds_regional}
\alias{get_box_bounds_regional}
\title{Get Excel box bounds}
\usage{
get_box_bounds_regional(
  sheet_name = c("Intro", "Indicator_view", "Country_view", "uhc_regional_sumamry",
    "hep_regional_summary", "hpop_regional_summary"),
  box = NULL
)
}
\arguments{
\item{sheet_name}{name of the sheet to get boundaries. It can be either:
\itemize{
\item Intro
\item Indicator_view
\item Country_view
\item uhc_regional_summary
\item hep_regional_summary
\item 
}}

\item{box}{name of which to get the boundaries. If NULL, all boxes for
\code{sheet_name} is returned.}
}
\value{
named list or list of list  with boxes bounds for specified
\code{sheet_name} and \code{box}. For each box, returns \code{start_row}, \code{end_row},
\code{start_col} and `end_col
}
\description{
\code{get_box_bounds_regional} gets the bounds (boundaries) of Excel regional
sheet boxes.
}
