% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{get_df_one_scenario}
\alias{get_df_one_scenario}
\title{Get data frame with just one scenario}
\usage{
get_df_one_scenario(df, scenario, default_scenario)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}
}
\value{
A data frame with one scenario
}
\description{
Get just one scenario in the data frame. Is used to avoid issues with when
multiple scenarios are present in export functions, as scenarios are not implemented
at the moment.
}
