% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{get_latest_reported_df}
\alias{get_latest_reported_df}
\title{Get latest reported data frame}
\usage{
get_latest_reported_df(
  df,
  value_col,
  transform_value_col = NULL,
  level = NULL,
  ind_df
)
}
\arguments{
\item{df}{Data frame in long format, where 1 row corresponds to a specific
country, year, and indicator.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{level}{integer indicating the level column. Specific to HEP.}

\item{ind_df}{data frame containing information on indicators}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{get_latest_reported()} gets the latest reported data available for estimated
or reported data. Used in write functions.
}
