% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison_indicator.R
\name{plot_comparison_indicator}
\alias{plot_comparison_indicator}
\title{Export plot comparison of two billionaiRe data frames for indicator}
\usage{
plot_comparison_indicator(
  ...,
  indicator,
  ind = "ind",
  iso3_col = "iso3",
  year_col = "year",
  value = "value",
  type_col = "type",
  ind_ids = billionaiRe::billion_ind_codes("all")
)
}
\arguments{
\item{...}{Name-value pair of data frames to use e.g. (new = df_1).The name
(if provided) gives the name of the data frame and the value is the data
frame itself. The name is used to appear in the legend of the plot if
provided.}

\item{indicator}{name of indicator to plot}

\item{ind}{Column name of column with indicator names.}

\item{iso3_col}{Column name of column with country ISO3 codes.}

\item{year_col}{Column name of column with years.}

\item{value}{Column name of column(s) with indicator's values, used to
calculate contributions.}

\item{type_col}{Column name of column with type data.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
a \code{ggplot2} object
}
\description{
\code{write_hep_summary_sheet()} writes all the content and styling for the HEP
summary sheet. Used within \code{export_hep_country_summary_xls()}
}
