% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_indicator.R
\name{plot_timeseries_indicator}
\alias{plot_timeseries_indicator}
\title{Plot timeseries for indicator}
\usage{
plot_timeseries_indicator(
  df,
  iso3,
  indicator,
  ind = "ind",
  iso3_col = "iso3",
  year_col = "year",
  transform_value = "transform_value",
  type_col = "type",
  scenario = "scenario",
  default_scenario = "default",
  start_year = 2018
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{iso3}{vector of ISO3 country codes to plot}

\item{indicator}{name of indicator to plot}

\item{ind}{Column name of column with indicator names.}

\item{iso3_col}{Column name of column with country ISO3 codes.}

\item{year_col}{Column name of column with years.}

\item{transform_value}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{type_col}{Column name of column with type data.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{default_scenario}{name of the default scenario.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}
}
\value{
list with two \code{ggplot2} objects: one with a common scale for all iso3
and one with an individual scale for each iso3.
}
\description{
Plot timeseries to provide line plots for each scenario present in \code{scenario}
with some manipulations. Produce two faceted plots where each facet is an
\code{iso3}.
}
