% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_style.R
\name{scenarios_style}
\alias{scenarios_style}
\title{Change font of scenario values}
\usage{
scenarios_style(
  df,
  wb,
  sheet_name,
  start_row,
  start_col,
  ind,
  year,
  type_col,
  scenario,
  ind_df,
  scenarios_order
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column with type data.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{scenarios_order}{(character) vector in which \code{scenario} should be \code{dplyr::arrange}}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{scenarios_style()} changes the font of time series values based on the
indicator type it is:
}
\details{
\itemize{
\item bold: reported
\item normal: estimated
\item faded: imputed or projected
}
}
