% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scenarios_sheet.R
\name{style_scenarios_sheet}
\alias{style_scenarios_sheet}
\title{Styles scenarios sheet}
\usage{
style_scenarios_sheet(
  df,
  wb,
  billion = c("hep", "hpop", "uhc"),
  sheet_name,
  start_row,
  start_col,
  scenario_col = "scenario",
  scenarios_order,
  df_wide,
  ind_df,
  ind_ids,
  this_iso3
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{billion}{Billion to be used for billion styling when relevant: either
"hep", "hpop", or "uhc" when no billion to be applied.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{scenarios_order}{(character) vector in which \code{scenario_col} should be \code{dplyr::arrange}}

\item{df_wide}{wide version of \code{df} generated by \code{write_uhc_timeseries_sheet}}

\item{ind_df}{data frame containing information on indicators}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{this_iso3}{character iditenfying the country being styled.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{style_header_hpop_summary_sheet} styles the title and sub-title of the worksheet header.
}
