% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_hpop_summary.R
\name{write_baseline_projection_hpop_summary}
\alias{write_baseline_projection_hpop_summary}
\title{Write the baseline/projection data frame to the data sheet}
\usage{
write_baseline_projection_hpop_summary(
  df,
  wb,
  sheet_name,
  ind_df,
  start_year,
  end_year,
  value_col,
  transform_value_col,
  bounds,
  ind_ids
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\description{
Write the baseline/projection data frame to the data sheet
}
