% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_hpop_summary.R
\name{write_billion_contribution_hpop_summary}
\alias{write_billion_contribution_hpop_summary}
\title{Write the overall contribution to billion to the data sheet}
\usage{
write_billion_contribution_hpop_summary(
  df,
  wb,
  sheet_name,
  contribution,
  contribution_pct,
  ind,
  year,
  end_year,
  bounds,
  iso,
  boxes_bounds
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value}.}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to
2025.}

\item{bounds}{named integer vector identifying integer sheet start and end reference column.
All parameters must be numerical. For converting Excel column references to integer references
see \code{\link[openxlsx:int2col]{openxlsx::int2col()}}.
Must have the following named elements:
\itemize{
\item \code{start_col}: start column
\item \code{end_col}: end column
\item \code{start_row}: start row
\item \code{end_row}: end row
}}

\item{iso}{ISO3 code of country to summarize.}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}
}
\description{
Used within \code{write_hpop_summary_sheet()}
}
