% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_hep_summary.R
\name{write_data_boxes_hep_summary}
\alias{write_data_boxes_hep_summary}
\title{Write and style data boxes in HEP summary data}
\usage{
write_data_boxes_hep_summary(
  df,
  pillar = c("prepare", "prevent", "detect_respond"),
  wb,
  sheet_name,
  value_col,
  transform_value_col,
  boxes_bounds,
  start_year,
  end_year,
  scenario_col,
  ind_df,
  ind_ids
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{pillar}{character Pillar identifying the pillar to style. Must be one of
\itemize{
\item RMNCH
\item infectious
\item NCD
\item service
}}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{boxes_bounds}{named list of bounds for data frame boxes to be written in sheet.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to
2025.}

\item{scenario_col}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_data_boxes_hep_summary} writes and styles the data box in UHC summary
worksheet. Used in \code{write_hep_summary_sheet()}.
}
