% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hpop_summary_sheet.R
\name{write_hpop_summary_sheet}
\alias{write_hpop_summary_sheet}
\title{Write HPOP summary sheet}
\usage{
write_hpop_summary_sheet(
  df,
  wb,
  sheet_name,
  iso,
  start_year = 2018,
  end_year = 2019:2025,
  value_col = "value",
  transform_value_col = "transform_value",
  contribution = "contribution",
  contribution_pct = paste0(contribution, "_percent"),
  contribution_pct_total_pop = paste0(contribution, "_percent_total_pop"),
  ind_df,
  ind_ids
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{sheet_name}{character name of the sheet to update}

\item{iso}{ISO3 code of country to summarize.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to
2025.}

\item{value_col}{Column name of column with indicator values.}

\item{transform_value_col}{Column name of column(s) with transformed indicator
values, used to calculate contributions.}

\item{contribution}{Column name of column(s) to store contribution (population)
values. Must be the same length as \code{transform_value_col}.}

\item{contribution_pct}{Column name of column(s) to store contribution (percent)
values. Must be the same length as \code{transform_value_col}.}

\item{contribution_pct_total_pop}{Column name of column(s) to store contribution
(percent of total population of the country) values. Must be the same length
as \code{transform_value_col}.}

\item{ind_df}{data frame containing the indicators in the correct order and format to be used.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{write_hpop_summary_sheet()} writes all the content and styling for the HPOP
summary sheet. Used within \code{export_hpop_country_summary_xls()}
}
