% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_regional_permanent_sheets.R
\name{write_regional_permanent_sheets}
\alias{write_regional_permanent_sheets}
\title{Write permanent sheets for regional summary}
\usage{
write_regional_permanent_sheets(bounds, start_year, end_year)
}
\arguments{
\item{bounds}{a named list with the extent of the data to be written over.
Must contain the following named parameters:
\itemize{
\item start_row
\item end_row
\item start_col
\item end_col
}}
}
\value{
A \code{openxlsx} workbook object
}
\description{
\code{write_regional_permanent_sheets()} writes and styles the Intro, Indicator
View, and Country View sheets of the regional summary workbook. Those sheets
are present in all cases taken into account by the different functions.
}
