% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scenarios_sheet.R
\name{write_scenario_sheet}
\alias{write_scenario_sheet}
\title{Write scenario sheet}
\usage{
write_scenario_sheet(
  df,
  wb = wb,
  billion = c("hep", "uhc", "hpop"),
  sheet_name = "Scenarios",
  start_row = 4,
  start_col = 2,
  value = "value",
  scenario = "scenario",
  ind_df,
  ind = "ind",
  year = "year",
  type_col = "type",
  ind_ids,
  start_year = 2018,
  end_year = 2025,
  iso3 = "iso3",
  default_scenario = "default"
)
}
\arguments{
\item{df}{data frame to be written}

\item{wb}{a \code{openxlsx} workbook to be edited.}

\item{billion}{Billion to be used for billion styling when relevant: either
"hep", "hpop", or "uhc" when no billion to be applied.}

\item{sheet_name}{character name of the sheet to update}

\item{start_row}{integer start row of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{start_col}{integer start col of styling. Passed to \code{\link[openxlsx:addStyle]{openxlsx::addStyle()}}}

\item{value}{Column name of column with indicator values.}

\item{scenario}{Column name of column with scenario identifiers. Useful for
calculating contributions on data in long format rather than wide format.}

\item{ind_df}{data frame containing information on indicators}

\item{ind}{Column name of column with indicator names.}

\item{year}{Column name of column with years.}

\item{type_col}{Column name of column with type data.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to 2025.}

\item{iso3}{Column name of column with country ISO3 codes.}

\item{default_scenario}{name of the default scenario.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_hpop_country_summary_xls} Export a country-specific for HPOP billion.
}
