% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot_timeseries.R
\name{export_plot_timeseries_indicator_pdf}
\alias{export_plot_timeseries_indicator_pdf}
\title{Export plot timeseries for an indicator to pdf}
\usage{
export_plot_timeseries_indicator_pdf(
  df,
  indicator,
  scale = c("combined", "free", "fixed", "free_x", "free_y"),
  output_folder = "outputs",
  experiment = NULL,
  group_iso3 = TRUE,
  default_scenario = "default",
  value_col = "value",
  start_year = 2018,
  ind_ids = billionaiRe::billion_ind_codes("all", include_calculated = TRUE),
  version = whdh::get_formatted_timestamp()
)
}
\arguments{
\item{df}{Data frame in long format filtered for a specific country, where 1 row corresponds
to a specific year, and indicator.}

\item{indicator}{name of indicator to plot}

\item{scale}{type of scale to be exported. Can be either:
\itemize{
\item \code{common}: same scale for all iso3
\item \code{free}: individual scale for each iso3
\item \code{combined}: combination of common and free.
By default, \code{combined} is used
}}

\item{output_folder}{Folder path to where the Excel files should be written}

\item{experiment}{name of the experiment of plots. Default to NULL. If used, a new
folder will be added to \code{output_folder} to store the results.}

\item{group_iso3}{if TRUE (default), iso3s will be grouped by 54 to have more readable graphs.}

\item{default_scenario}{name of the default scenario.}

\item{value_col}{Column name of column with indicator values.}

\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{ind_ids}{Named vector of indicator codes for input indicators to the Billion.
Although separate indicator codes can be used than the standard, they must
be supplied as a named vector where the names correspond to the output of
\code{billion_ind_codes()}.}

\item{version}{character vector identifying the version of the plot to be passed.}
}
\value{
\code{openxslx} Workbook object. Output file is in \code{output_folder}.
}
\description{
\code{export_plot_timeseries_indicator_pdf} exports plots to pdf.
}
