% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_regional_permanent_sheets.R
\name{write_regional_permanent_sheets}
\alias{write_regional_permanent_sheets}
\title{Write permanent sheets for regional summary}
\usage{
write_regional_permanent_sheets(start_year, end_year)
}
\arguments{
\item{start_year}{Base year for contribution calculation, defaults to 2018.}

\item{end_year}{End year(s) for contribution calculation, defaults to 2019 to
2025.}
}
\value{
A \code{openxlsx} workbook object
}
\description{
\code{write_regional_permanent_sheets()} writes and styles the Intro, Indicator
View, and Country View sheets of the regional summary workbook. Those sheets
are present in all cases taken into account by the different functions.
}
